Experimental Debian 10 port

+++ AppKit/X11.subproj/KTFont_FT.h
@@ -12,7 +12,7 @@
 #import <ft2build.h>
 
 #ifdef LINUX
-typedef int ptrdiff_t;
+typedef long ptrdiff_t;
 #endif
 #import FT_FREETYPE_H
 #import FT_RENDER_H
